fixed bug in IoJoinThread() added IoCreateClientByConnect() fixed warning in Utf8ToUtf16() and AnsiToUtf16()

+++ trunk/util/strconv.c
@@ -17,7 +17,7 @@ CSTATUS Utf8ToUtf16(
 	size_t BufferSizeInChars,
 	const char *Input)
 {
-	int result = MultiByteToWideChar(CP_UTF8, 0, Input, -1, Buffer, BufferSizeInChars);
+	int result = MultiByteToWideChar(CP_UTF8, 0, Input, -1, Buffer, (int)BufferSizeInChars);
 	if (result == 0)
 		return Win32ErrorCodeToCStatus(GetLastError());
 	return C_SUCCESS;
@@ -39,7 +39,7 @@ CSTATUS AnsiToUtf16(
 	size_t BufferSizeInChars,
 	const char *Input)
 {
-	int result = MultiByteToWideChar(CP_ACP, 0, Input, -1, Buffer, BufferSizeInChars);
+	int result = MultiByteToWideChar(CP_ACP, 0, Input, -1, Buffer, (int)BufferSizeInChars);
 	if (result == 0)
 		return Win32ErrorCodeToCStatus(GetLastError());
 	return C_SUCCESS;
